Terrain Based Grid Generation
Adapted grid plane generator into one that wraps and sizes to a Unity Terrain and adjusts to the height based on each vertices location. This version does not correctly size to terrains with fractional widths, and does not currently allow for differing grid tile sizes.

Additional changes include the addition of error handling.